Added the assert statements

+++ employee_user/src/main/java/com/m_landalex/employee_user/service/AddressService.java
@@ -35,7 +35,9 @@ public class AddressService {
 			rabbitTemplate.convertAndSend(queueName, "error");
 			throw new AsyncXAResourcesException("Address object is null -> method save");
 		}
+		assert address != null : "AddressService.class, method save address is null";
 		Address newAddress = mapper.toObject(repository.save(mapper.toEntity(address)));
+		assert newAddress != null : "AddressService.class, method save newAddress is null";
 		rabbitTemplate.convertAndSend(queueName, "succesful");
 		return newAddress;
 	}
@@ -43,6 +45,7 @@ public class AddressService {
 	@Transactional(readOnly = true)
 	public Collection<Address> fetchAll() {
 		Optional<Collection<Address>> optional = Optional.of(mapper.toObjectList(repository.findAll()));
+		assert optional != null : "AddressService.class, method fetchAll optional is null";
 		if(optional.isPresent()) {
 			return optional.stream().flatMap(collection -> collection.stream()).collect(Collectors.toList());
 		}
@@ -51,6 +54,7 @@ public class AddressService {
 
 	@Transactional(readOnly = true)
 	public Address fetchById(Long id) {
+		assert id != null : "AddressService.class, method fetchById id is null";
 		Optional<Address> optional = Optional.of(mapper.toObject(repository.findById(id).get()));
 		if(optional.isPresent()) {
 			return optional.get();
@@ -62,7 +66,10 @@ public class AddressService {
 	 * only addresses that are not related to any are employee
 	 */
 	public void deleteById(Long id) {
-		repository.delete(repository.findById(id).get());
+		assert id != null : "AddressService.class, method deleteById id is null";
+		if(id != null) {
+			repository.delete(repository.findById(id).get());
+		}
 	}
 	
 	@Transactional(propagation = Propagation.NEVER)
@@ -72,6 +79,7 @@ public class AddressService {
 	
 	@Transactional(readOnly = true)
 	public Address fetchByCity(String city) {
+		assert city != null : "AddressService.class, method fetchByCity city is null";
 		Optional<Address> optional = Optional.of(mapper.toObject(repository.findByCity(city)));
 		if(optional.isPresent()) {
 			if(optional.stream().allMatch(address -> address.getCity().compareToIgnoreCase(city) == 0)) {
